# openapi.yaml
openapi: 3.0.3
info:
  title: DOGFOOD API
  version: 0.1.0
  description: 'Hackathon judging, voting, and management API. Note that admin/service routes typically wrap responses in { data: ... } while some public routes return unwrapped JSON. Error formats also vary between wrapped structured objects and raw string messages.'
servers:
  - url: 'http://localhost:3000'
    description: Local deployment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorStandard:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Hackathon:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - COMPLETED
        policy_id:
          type: string
          format: uuid
          nullable: true
        required_judges:
          type: integer
          nullable: true
        submissions_close:
          type: string
          format: date-time
          nullable: true
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    GallerySubmission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        team_name:
          type: string
...
    AggregateResult:
      type: object
      properties:
        submission_id:
          type: string
          format: uuid
        title:
          type: string
        team_id:
          type: string
          format: uuid
        total_assignments:
          type: integer
        submitted_evaluations:
          type: integer
        missing_evaluations:
          type: integer
        aggregate_score:
          type: number
          nullable: true
        status:
          type: string
    PublicResult:
      type: object
      properties:
        submission_id:
          type: string
          format: uuid
        title:
          type: string
        team_name:
          type: string
        aggregate_score:
          type: number
          nullable: true
        community_votes:
          type: integer
...
paths:
...
  /api/hackathons:
    get:
      summary: List hackathons
      description: Get a list of all hackathons. Requires Organizer access.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of hackathons
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Hackathon'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
        '403':
          description: Forbidden (Organizer required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      operationId: getapihackathons
    post:
      summary: Create hackathon
      description: Create a new hackathon. Requires Organizer access.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 10000
                  default: ''
      responses:
        '201':
          description: Hackathon created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Hackathon'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
        '403':
          description: Forbidden (Organizer required)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorStandard'
      operationId: postapihackathons
...
  '/api/hackathons/{id}/results':
    get:
      summary: Get hackathon results (Admin)
      description: Retrieve complete scoring aggregation. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Results aggregation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/AggregateResult'
      operationId: getapihackathonsidresults
...
  /api/public/gallery:
    get:
      summary: Public submission gallery
      description: Returns a list of all hackathon submissions. Returns raw JSON array without a wrapper.
      parameters:
        - in: query
          name: seed
          schema:
            type: string
          description: Optional seed for deterministic random shuffling
      responses:
        '200':
          description: Array of gallery submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GallerySubmission'
      security: []
      operationId: getapipublicgallery
...
  '/api/public/hackathons/{hackathonId}/results':
    get:
      summary: Public hackathon results
      description: Returns the public results summary for a hackathon. Enforces results embargo if submissions are not yet closed.
      parameters:
        - in: path
          name: hackathonId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Public results overview
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/PublicResult'
        '403':
          description: Results embargoed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      security: []
      operationId: getapipublichackathonshackathonIdresults
  /api/public/submit:
    post:
      summary: Public create submission
      description: Participant endpoint to create a submission. Requires authentication.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                summary:
                  type: string
      responses:
        '201':
          description: Submission created successfully (Unwrapped payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
      operationId: postapipublicsubmit
  '/api/public/submissions/{submissionId}':
    put:
      summary: Edit public submission
      description: Modifies an existing submission. Enforces ownership so only team members can edit.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
      responses:
        '200':
          description: Submission updated (Unwrapped payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '403':
          description: Not authorized to edit (Not team member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      operationId: putapipublicsubmissionssubmissionId
  '/api/public/submissions/{submissionId}/vote':
    post:
      summary: Cast community vote
      description: 'Authenticated user votes for a submission. Rejects self-voting, duplicate voting, and voting past deadline. Returns {error: string} on failure.'
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hackathon_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Vote recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CommunityVote'
        '400':
          description: Invalid Request Format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '403':
          description: Self-vote or Voting closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '409':
          description: Duplicate vote
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      operationId: postapipublicsubmissionssubmissionIdvote
  '/api/public/submissions/{submissionId}/votes':
    get:
      summary: Get vote count
      description: Returns the total community votes for a submission.
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Vote count total
          content:
            application/json:
              schema:
                type: object
                properties:
                  submission_id:
                    type: string
                    format: uuid
                  votes:
                    type: integer
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      security: []
      operationId: getapipublicsubmissionssubmissionIdvotes
  '/api/public/submissions/{submissionId}/comments':
    get:
      summary: Get comments
      description: List public comments on a submission. Returns unwrapped JSON array.
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Array of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      security: []
      operationId: getapipublicsubmissionssubmissionIdcomments
    post:
      summary: Create comment
      description: Authenticated user creates a comment on a submission.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  maxLength: 2000
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '403':
          description: Commenting closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
        '404':
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      operationId: postapipublicsubmissionssubmissionIdcomments
...
  '/api/public/records/{recordId}':
    get:
      summary: Get a verifiable record
      description: 'Fetches the payload, cryptographic signature, and key ID (kid) for a public verification record.'
      parameters:
        - in: path
          name: recordId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A verified record
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      type:
                        type: string
                        enum:
                          - JUDGE_PARTICIPATION
                          - PARTICIPANT_CERTIFICATE
                      payload:
                        type: object
                        additionalProperties: true
                      signature:
                        type: string
                      kid:
                        type: string
                      issued_at:
                        type: string
                        format: date-time
        '404':
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorString'
      security: []
      operationId: getapipublicrecordsrecordId
...